h1de/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.0.1
numpy==1.25.2
python-dateutil==2.8.2